ges_student_project/requirements.txt
django == 4.2
djangorestframework == 3.15.2
djangorestframework-simplejwt == 5.4.0
django-cors-headers == 4.7.0